requirements.txt
Flask-HTTPAuth==3.2.1
click==6.6
Flask==0.11.1
Flask-OAuthlib==0.9.3
GoogleAppEngineCloudStorageClient==1.9.22.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==2.0.1
passlib==1.7.0
requests==2.12.4
requests-oauthlib==0.7.0
Werkzeug==0.11.11
